places: move places page and sections into panel-plugin/places/
