+++ panel-plugin/meson.build
@@ -8,10 +8,10 @@ plugin_sources = [
   'command-edit.cpp',
   'element.cpp',
   'favorites-page.cpp',
-  'favourites-section.cpp',
-  'history-section.cpp',
-  'home-search-worker.cpp',
-  'home-section.cpp',
+  'places/favourites-section.cpp',
+  'places/history-section.cpp',
+  'places/home-search-worker.cpp',
+  'places/home-section.cpp',
   'icon-renderer.cpp',
   'icon-size.cpp',
   'image-menu-item.h',
@@ -20,8 +20,8 @@ plugin_sources = [
   'launcher-tree-view.cpp',
   'migration.cpp',
   'page.cpp',
-  'places-item.cpp',
-  'places-page.cpp',
+  'places/places-item.cpp',
+  'places/places-page.cpp',
   'plugin.cpp',
   'profile.cpp',
   'query.cpp',
