+++ panel-plugin/places/favourites-section.cpp
