+++ panel-plugin/places/favourites-section.h
