+++ panel-plugin/places/history-section.cpp
