+++ panel-plugin/places/history-section.h
