+++ panel-plugin/places/home-search-worker.cpp
