+++ panel-plugin/places/home-search-worker.h
