+++ panel-plugin/places/home-section.cpp
