+++ panel-plugin/places/home-section.h
