+++ panel-plugin/places/places-item.cpp
