+++ panel-plugin/places/places-item.h
