+++ panel-plugin/places/places-page.cpp
