+++ panel-plugin/places/places-section.h
