+++ panel-plugin/window.cpp
@@ -21,11 +21,11 @@
 #include "launcher/category-button.h"
 #include "command.h"
 #include "favorites-page.h"
-#include "favourites-section.h"
-#include "history-section.h"
-#include "home-section.h"
+#include "places/favourites-section.h"
+#include "places/history-section.h"
+#include "places/home-section.h"
 #include "launcher-view.h"
-#include "places-page.h"
+#include "places/places-page.h"
 #include "plugin.h"
 #include "profile.h"
 #include "recent-page.h"
