+++ po/POTFILES
@@ -5,17 +5,17 @@ panel-plugin/command-edit.cpp
 panel-plugin/command.cpp
 panel-plugin/element.cpp
 panel-plugin/favorites-page.cpp
-panel-plugin/favourites-section.cpp
-panel-plugin/history-section.cpp
-panel-plugin/home-section.cpp
+panel-plugin/places/favourites-section.cpp
+panel-plugin/places/history-section.cpp
+panel-plugin/places/home-section.cpp
 panel-plugin/icon-renderer.cpp
 panel-plugin/icon-size.cpp
 panel-plugin/launcher-icon-view.cpp
 panel-plugin/launcher-tree-view.cpp
 panel-plugin/launcher.cpp
 panel-plugin/page.cpp
-panel-plugin/places-item.cpp
-panel-plugin/places-page.cpp
+panel-plugin/places/places-item.cpp
+panel-plugin/places/places-page.cpp
 panel-plugin/plugin.cpp
 panel-plugin/preset.cpp
 panel-plugin/profile.cpp
